requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
trimesh==4.0.5
rembg
imageio[ffmpeg]
gradio
xatlas==0.0.9
moderngl==5.10.0
huggingface_hub==0.24.6
transformers==4.44.2
onnxruntime
pyrender
torch
torchvision
torchaudio
numpy==1.26.4
